Refactor proxy to use startDebugging
